rearranging components
